chore(coin-tezos): clean tezos module for coin-modules migration
