+++ libs/coin-modules/coin-tezos/package.json
@@ -28,12 +28,8 @@
       "lib-es/*": [
         "lib-es/*"
       ],
-      "specs": [
-        "lib/test/bot-specs"
-      ],
       "*": [
         "lib/*",
-        "lib/bridge/*",
         "lib/test/*",
         "lib/types/*"
       ]
@@ -42,25 +38,10 @@
   "exports": {
     "./lib/*": "./lib/*.js",
     "./lib-es/*": "./lib-es/*.js",
-    "./deviceTransactionConfig": {
-      "@ledgerhq/source": "./src/deviceTransactionConfig.ts",
-      "require": "./lib/deviceTransactionConfig.js",
-      "default": "./lib-es/deviceTransactionConfig.js"
-    },
     "./errors": {
       "@ledgerhq/source": "./src/types/errors.ts",
       "require": "./lib/types/errors.js",
       "default": "./lib-es/types/errors.js"
-    },
-    "./specs": {
-      "@ledgerhq/source": "./src/test/bot-specs.ts",
-      "require": "./lib/test/bot-specs.js",
-      "default": "./lib-es/test/bot-specs.js"
-    },
-    "./transaction": {
-      "@ledgerhq/source": "./src/transaction.ts",
-      "require": "./lib/transaction.js",
-      "default": "./lib-es/transaction.js"
     },
     "./*": {
       "@ledgerhq/source": "./src/*.ts",
@@ -72,13 +53,11 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@ledgerhq/coin-module-framework": "catalog:",
-    "@ledgerhq/cryptoassets": "workspace:^",
     "@ledgerhq/devices": "workspace:*",
     "@ledgerhq/errors": "workspace:^",
     "@ledgerhq/ledger-wallet-framework": "workspace:^",
     "@ledgerhq/live-network": "workspace:^",
     "@ledgerhq/logs": "workspace:^",
-    "@ledgerhq/types-live": "workspace:^",
     "@taquito/ledger-signer": "^23.0.0-RC.0",
     "@taquito/rpc": "^23.0.0-RC.0",
     "@taquito/taquito": "^23.0.0-RC.0",
@@ -90,7 +69,6 @@
   },
   "devDependencies": {
     "@ledgerhq/disable-network-setup": "workspace:^",
-    "@ledgerhq/types-cryptoassets": "workspace:^",
     "@swc/core": "catalog:",
     "@swc/jest": "catalog:",
     "@taquito/local-forging": "^23.0.0-RC.0",
