+++ libs/coin-modules/coin-tezos/src/index.ts
@@ -1,3 +1,1 @@
 export * from "./types";
-export * from "./transaction";
-export * from "./serialization";
