+++ libs/coin-modules/coin-tezos/src/types/index.ts
@@ -1,4 +1,2 @@
-export * from "./bridge";
 export * from "./errors";
 export * from "./model";
-export * from "./signer";
