+++ libs/ledger-live-common/src/bridge/generic-coin-framework/families/tezos/accountRawAssign.ts
@@ -1,4 +1,4 @@
-import { assignFromAccountRaw, assignToAccountRaw } from "@ledgerhq/coin-tezos/serialization";
+import { assignFromAccountRaw, assignToAccountRaw } from "../../../../families/tezos/serialization";
 
 /**
  * Tezos-specific hooks that persist `stakingPositions` through the
