+++ libs/ledger-live-common/src/coin-modules/transaction-types.ts
@@ -147,7 +147,7 @@ import type {
   TransactionRaw as tezosTransactionRaw,
   TransactionStatus as tezosTransactionStatus,
   TransactionStatusRaw as tezosTransactionStatusRaw,
-} from "@ledgerhq/coin-tezos/types/index";
+} from "../families/tezos/types";
 import type {
   Transaction as tonTransaction,
   TransactionRaw as tonTransactionRaw,
