+++ libs/ledger-live-common/src/families/tezos/deviceTransactionConfig.ts
