+++ libs/ledger-live-common/src/families/tezos/hw-signMessage.ts
@@ -1,7 +1,7 @@
 import { SignerContext } from "@ledgerhq/ledger-wallet-framework/signer";
 import { Account, AnyMessage, DeviceId } from "@ledgerhq/types-live";
-import { Curve, TezosSigner } from "./types/signer";
-import { convertSecp256k1DERToRaw } from "./utils";
+import type { Curve, TezosSigner } from "./types";
+import { convertSecp256k1DERToRaw } from "@ledgerhq/coin-tezos/utils";
 
 // Curve tag (matching hw-app-tezos `TezosCurves`) from the implicit-address prefix.
 function curveForAddress(address: string): Curve {
