+++ libs/ledger-live-common/src/families/tezos/serialization.ts
@@ -1,12 +1,7 @@
 import type { Account, AccountRaw } from "@ledgerhq/types-live";
 import { log } from "@ledgerhq/logs";
 import { BigNumber } from "bignumber.js";
-import type {
-  StakingPosition,
-  StakingPositionRaw,
-  TezosAccount,
-  TezosAccountRaw,
-} from "./types/bridge";
+import type { StakingPosition, StakingPositionRaw, TezosAccount, TezosAccountRaw } from "./types";
 
 function safeISOString(d: Date, uid: string): string | undefined {
   if (Number.isFinite(d.getTime())) return d.toISOString();
