+++ libs/ledger-live-common/src/families/tezos/staking.ts
@@ -1,2 +1,2 @@
-export { getAccountDelegationSync, isAccountDelegating } from "@ledgerhq/coin-tezos/network/bakers";
+export { getAccountDelegationSync, isAccountDelegating } from "./bakers";
 export { default as whitelist } from "@ledgerhq/coin-tezos/network/bakers.whitelist-default";
